Move pollution data routines to a separate module
