--- water/__init__.py
@@ -1,5 +1,0 @@
-from .main import pollution_series
-
-__all__ = [
-    'pollution_series'
-]
